change name from serve_ts to serve_typescript
